requirements.txt
numpy==1.24.3
streamlit>=1.44.0
torch>=1.9.0
torchvision==0.15.2
librosa>=0.9.1
matplotlib>=3.4.3
pillow>=8.3.2
soundfile>=0.10.3
opencv-python-headless
scikit-learn>=1.0.0
streamlit-webrtc==0.47.1
stripe>=5.4.0
PyJWT>=2.6.0
pandas>=1.5.0
toml>=0.10.2
